Adjust import in test script according to DataPlugin name
